Move components to own directory
