+++ src/App.js
@@ -1,16 +1,5 @@
 import { useState } from 'react';
-
-function Square({ val, fooClick, winner }) {
-
-  return(
-    <button
-      className={`square ${winner}`}
-      onClick={fooClick}
-    >
-      { val }
-    </button>
-  );
-}
+import Board from './components/board.js';
 
 /* React words of wisdom: 
  * To collect data from multiple children, or to have two child components communicate with each other,
@@ -18,55 +7,6 @@ function Square({ val, fooClick, winner }) {
  * back down to the children via props. This keeps the child components in sync with each other and with their parent.
  *
  * Lifting state into a parent component is common when React components are refactored. */
-
-function Board({ xActive, spaces, onPlay }) {
-  let boardSize = 3;
-  let count = 0;
-
-  function barClick(pos) {
-    if(spaces[pos] || calculateWinner(spaces)) return;
-
-    let newSpaces = spaces.slice();
-
-    let newCharacter = xActive ? 'X' : 'O'
-    newSpaces[pos] = newCharacter;
-    onPlay(newSpaces);
-  }
-
-  let winners = calculateWinner(spaces);
-
-  if(winners) {
-    status = `${xActive ? 'O' : 'X'} is our winner!`;
-  } else if(!spaces.includes(null)) {
-    status = "It's a draw!";
-  } else {
-    status = `It is ${xActive ? 'X' : 'O'}'s turn`
-  }
-
-   const renderRow = (rowIndex) => (
-     <div className="board-row" key={rowIndex}>
-       {[0, 1, 2].map((colIndex) => {
-         const position = rowIndex * boardSize + colIndex;
-         let winning_square = winners?.includes(position);
-         return (
-           <Square
-             key={colIndex}
-             val={spaces[position]}
-             winner={ winning_square ? "winner" : ""}
-             fooClick={() => barClick(position)}
-           />
-         );
-       })}
-     </div>
-   );
-  
-  return (
-    <>
-      <div className="status">{status}</div>
-      {[0, 1, 2].map((rowIndex) => renderRow(rowIndex))}
-    </>
-  );
-}
 
 export default function Game() {
   const [xActive, setXActive] = useState(true);
@@ -116,26 +56,3 @@ export default function Game() {
     </div>
   );
 }
-
-// We'll just used the supplied alg for now
-// -- Changing things up to return the winning row
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return lines[i];
-    }
-  }
-  return null;
-}
-
