+++ src/components/square.js
@@ -1,0 +1,10 @@
+export default function Square({ val, fooClick, winner }){
+  return(
+    <button
+      className={`square ${winner}`}
+      onClick={fooClick}
+    >
+      { val }
+    </button>
+  );
+}
